Move world logic
